backend/requirements.txt
algorithmia==1.0.8
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cssselect==1.0.3
dj-database-url==0.4.2
Django==2.0
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-guardian==1.4.9
django-jet==1.0.6
djangorestframework==3.7.7
enum34==1.1.6
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.0
parsel==1.4.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pytest==3.2.3
pytz==2018.3
queuelib==1.4.2
raven==6.5.0
requests==2.18.4
scipy==1.0.0
Scrapy==1.5.0
selenium==3.6.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
uritemplate==3.0.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
